Make FLASH_CCA writing as optional.
Currently the linker script always generates .flash_cca section
at the end of flash. However this section is very rarely going to be
changed. After the first time programming, it is going to be the
for most of the life. So make CCA flashing as optional. In other
words the linker wont generate .flash_cca section so the BIN
file would be much smaller.