Added make rules to create p[1|2].[elf|bin].
For OTA, we need to split the firmware into two halves.
When firmware is running from one half the half can be
overwritten with new firmware. Once done the CCA will
be updated and the board will be restarted. For this
we need to create two firmwares with different load
address(since we dont have dynamic loader).