OTA firmware update.
By splitting the flash memory into two portions(active and inactive),
firmware update can be implemented for CC2538. For this two
firmware has to be generated one with linked at 0x200000 and another
one which linked at 0x240000. Based on the current running firmware,
the new firmware would be selected and updated in the unused portion.

To make it over the air(OTA), a CoAP resource(/debug/update) is
introduced. This resource will serve block PUT request and during
each incoming block it will write it directly to the flash.
Once all the blocks are written it will dynamically adjust the
CCA so that the next boot will bring up the new firmware.